removed unneeded parameter from JdbcEventStore
(LLM Generated:) JDBC version of the EventStoreTest

Note that it's a pain to reuse the EventStoreTest for the JDBC, because it needs Spring, and I don't want it to be mixed up with pure-I/O-free test. Unfortunately, due to the nested classes, it's not easy to instantiate the I/O-free EventStoreTest to reuse it. So, I simply asked the LLM to make JdbcEventStoreTest a copy of EventStoreTest.